Add files via upload
creates a multi conversion tool. The program must allow the user to choose the type of conversion that they require, perform the conversion and output the result of the conversion.  The user should then be able to perform another conversion.  The multi conversion tool will allow the user to convert from inches to feet (12 inches equals 1 foot), from pounds to Euro (1 Euro equals 0.79 pounds), or from Celsius to Fahrenheit (celsius = 5/9 (fahr – 32) or fahr = (celsius * 9/5) + 32)).